Remove legacy doc_store shim module
