+++ app/core/document_store/__init__.py
@@ -1,0 +1,157 @@
+"""Document store public API and shared structures."""
+
+from __future__ import annotations
+
+from dataclasses import dataclass, field
+from typing import TYPE_CHECKING, Any, List
+
+if TYPE_CHECKING:  # pragma: no cover - import for typing only
+    from ..model import Requirement
+
+
+class ValidationError(Exception):
+    """Raised when requirement links or payload violate business rules."""
+
+
+class RequirementError(Exception):
+    """Base class for requirement storage exceptions."""
+
+
+class DocumentNotFoundError(RequirementError):
+    """Raised when a document prefix is unknown."""
+
+    def __init__(self, prefix: str) -> None:
+        self.prefix = prefix
+        super().__init__(f"unknown document prefix: {prefix}")
+
+
+class RequirementNotFoundError(RequirementError):
+    """Raised when a requirement identifier cannot be located."""
+
+    def __init__(self, rid: str) -> None:
+        self.rid = rid
+        super().__init__(f"requirement {rid} not found")
+
+
+class RevisionMismatchError(RequirementError):
+    """Raised when the caller provides an outdated revision."""
+
+    def __init__(self, expected: int, actual: int) -> None:
+        self.expected = expected
+        self.actual = actual
+        super().__init__(f"revision mismatch: expected {expected}, have {actual}")
+
+
+@dataclass
+class LabelDef:
+    """Definition of a label available to document items."""
+
+    key: str
+    title: str
+    color: str | None = None
+
+
+@dataclass
+class DocumentLabels:
+    """Label configuration for a document."""
+
+    allow_freeform: bool = False
+    defs: List[LabelDef] = field(default_factory=list)
+
+
+@dataclass
+class Document:
+    """Configuration describing a document in the hierarchy."""
+
+    prefix: str
+    title: str
+    digits: int
+    parent: str | None = None
+    labels: DocumentLabels = field(default_factory=DocumentLabels)
+    attributes: dict[str, Any] = field(default_factory=dict)
+
+
+@dataclass
+class RequirementPage:
+    """Represent a paginated slice of requirements."""
+
+    items: list["Requirement"]
+    total: int
+    page: int
+    per_page: int
+
+
+from .documents import (  # noqa: E402
+    collect_label_defs,
+    collect_labels,
+    is_ancestor,
+    label_color,
+    load_document,
+    load_documents,
+    save_document,
+    stable_color,
+    validate_labels,
+)
+from .items import (  # noqa: E402
+    create_requirement,
+    delete_requirement,
+    get_requirement,
+    item_path,
+    list_item_ids,
+    list_requirements,
+    load_item,
+    next_item_id,
+    parse_rid,
+    patch_requirement,
+    rid_for,
+    save_item,
+    search_requirements,
+)
+from .links import (  # noqa: E402
+    delete_document,
+    delete_item,
+    iter_links,
+    link_requirements,
+    plan_delete_document,
+    plan_delete_item,
+)
+
+__all__ = [
+    "ValidationError",
+    "RequirementError",
+    "DocumentNotFoundError",
+    "RequirementNotFoundError",
+    "RevisionMismatchError",
+    "LabelDef",
+    "DocumentLabels",
+    "Document",
+    "RequirementPage",
+    "collect_label_defs",
+    "collect_labels",
+    "is_ancestor",
+    "label_color",
+    "load_document",
+    "load_documents",
+    "save_document",
+    "stable_color",
+    "validate_labels",
+    "create_requirement",
+    "delete_requirement",
+    "get_requirement",
+    "item_path",
+    "list_item_ids",
+    "list_requirements",
+    "load_item",
+    "next_item_id",
+    "parse_rid",
+    "patch_requirement",
+    "rid_for",
+    "save_item",
+    "search_requirements",
+    "delete_document",
+    "delete_item",
+    "iter_links",
+    "link_requirements",
+    "plan_delete_document",
+    "plan_delete_item",
+]
