+++ app/ui/main_frame.py
@@ -11,7 +11,7 @@ from ..agent import LocalAgent
 from ..config import ConfigManager
 from ..confirm import confirm
 from ..core.model import Requirement
-from ..core.doc_store import Document, LabelDef, save_document
+from ..core.document_store import Document, LabelDef, save_document
 from ..i18n import _
 from ..log import logger
 from ..mcp.controller import MCPController
