+++ tests/unit/test_cli_item.py
@@ -5,7 +5,7 @@ from pathlib import Path
 import pytest
 
 from app.cli import commands
-from app.core.doc_store import Document, save_document
+from app.core.document_store import Document, save_document
 
 
 @pytest.mark.unit
