+++ tests/unit/test_cli_item_add_labels.py
@@ -5,7 +5,7 @@ from pathlib import Path
 import pytest
 
 from app.cli import commands
-from app.core.doc_store import (
+from app.core.document_store import (
     Document,
     DocumentLabels,
     LabelDef,
