+++ tests/unit/test_cli_link.py
@@ -5,7 +5,7 @@ from pathlib import Path
 import pytest
 
 from app.cli import commands
-from app.core.doc_store import Document, save_document, save_item
+from app.core.document_store import Document, save_document, save_item
 
 
 @pytest.mark.unit
