+++ tests/unit/test_list_requirements_rid.py
@@ -1,4 +1,4 @@
-from app.core.doc_store import Document, save_document, save_item
+from app.core.document_store import Document, save_document, save_item
 from app.mcp.tools_read import list_requirements
